Update policy based DSA tmplt.cpp

+++ policy based DSA tmplt.cpp
@@ -1,7 +1,7 @@
 #include <ext/pb_ds/assoc_container.hpp>
 using namespace __gnu_pbds;
 
-typedef tree< ll, null_type, less_equal<ll>, rb_tree_tag, tree_order_statistics_node_update > ordered_set;
+typedef tree< long long, null_type, less_equal<long long>, rb_tree_tag,tree_order_statistics_node_update > ordered_set;
 /// change ll to any data type
 /// less_equal for multiset increasing order
 /// less for set increasing order
